TP2000-383 Add new workbasket pages from design (#623)
* Add workbasket icon and link to header bar

* Add confirm create page for workbaskets

* Add preview and review pages, update links

* Update buttons, content

* Update tests

* Move edit/preview/review workbasket functionality into workbasket app, and fix pagination

* Move DashboardView functionality into ReviewWorkbasketView

* Add permissions for workbasket views

* Amend templates and tests to reflect url change

* WIP: Remove Dashboard and move /delete tests

* Rename index to home

* Move over workbasket templates
