+++ common/jinja2/layouts/list_vertical.jinja
@@ -11,7 +11,7 @@
 {% block breadcrumb %}
   {{ govukBreadcrumbs({
     "items": [
-      {"text": "Home", "href": url("index")},
+      {"text": "Home", "href": url("home")},
       {"text": page_title}
     ]
   }) }}
