+++ common/static/common/scss/_components.scss
@@ -33,3 +33,13 @@
     text-decoration: underline;
   }
 }
+
+.govuk-header {
+
+  .workbasket-link {
+
+    svg {
+      margin-bottom: -4px;
+    }
+  }
+}
