+++ geo_areas/jinja2/geo_areas/detail.jinja
@@ -15,7 +15,7 @@
 {% block breadcrumb %}
   {{ govukBreadcrumbs({
     "items": [
-      {"text": "Home", "href": url("index")},
+      {"text": "Home", "href": url("home")},
       {"text": "Find and edit geographical areas", "href": url("geo_area-ui-list")},
       {"text": page_title}
     ]
