+++ importer/jinja2/importer/create.jinja
@@ -9,7 +9,7 @@
 {% block beforeContent %}
   {{ govukBreadcrumbs({
     "items": [
-      {"text": "Home", "href": url("index")},
+      {"text": "Home", "href": url("home")},
       {"text": "Find and edit import batches", "href": url("import_batch-ui-list")},
       {"text": page_title}
     ]
