+++ measures/jinja2/measures/create-start.jinja
@@ -14,7 +14,7 @@
 
   <p class="govuk-body">
     You'll also need to use the 
-      <a class="govuk-link" href="{{ url("index") }}">
+      <a class="govuk-link" href="{{ url("home") }}">
         manage trade tariffs service  
       </a>
     to find or create details of the following:
