+++ pii-ner-exclude.txt
@@ -26,6 +26,7 @@ ImportBatch
 BatchDependencies
 Jira
 MeasureActions
+do?{%
 MeasureAction
 Quota
 url('my
