+++ workbaskets/jinja2/workbaskets/confirm_create.jinja
@@ -1,0 +1,28 @@
+{% extends "layouts/layout.jinja" %}
+
+{% from "components/panel/macro.njk" import govukPanel %}
+{% from "components/button/macro.njk" import govukButton %}
+
+{% set page_title = "Create a new " ~ object._meta.verbose_name %}
+
+{% block beforeContent %}
+  {{ govukBreadcrumbs({
+    "items": [
+      {"text": "Home", "href": url("home")},
+      {"text": page_title}
+    ]
+  }) }}
+{% endblock %}
+
+{% block content %}
+  <div class="govuk-grid-row">
+    <div class="govuk-grid-column-two-thirds">
+      {{ govukPanel({
+        "titleText": object._meta.verbose_name|title ~ " " ~ object.id,
+        "text": "You have created a new workbasket",
+        "classes": "govuk-!-margin-bottom-7"
+      }) }}
+      <a href="{{ url('workbaskets:edit-workbasket') }}?workbasket={{ object.id }}" class="govuk-button govuk-button--primary">Continue to workbasket</a>
+    </div>
+  </div>
+{% endblock %}
