+++ workbaskets/jinja2/workbaskets/create.jinja
@@ -6,7 +6,7 @@
 {% block breadcrumb %}
     {{ govukBreadcrumbs({
       "items": [
-        {"text": "Home", "href": url("index")},
+        {"text": "Home", "href": url("home")},
         {"text": "Create a new workbasket"}
       ]
     }) }}
