+++ workbaskets/jinja2/workbaskets/delete_changes_confirm.jinja
@@ -9,7 +9,7 @@
 {% block breadcrumb %}
   {{ govukBreadcrumbs({
     "items": [
-      {"text": "Home", "href": url("index")},
+      {"text": "Home", "href": url("home")},
       {"text": page_title}
     ]
   }) }}
