+++ workbaskets/jinja2/workbaskets/detail.jinja
@@ -12,7 +12,7 @@
 {% block breadcrumb %}
   {{ govukBreadcrumbs({
     "items": [
-      {"text": "Home", "href": url("index")},
+      {"text": "Home", "href": url("home")},
       {"text": "Find and edit work baskets", "href": url("workbaskets:workbasket-ui-list")},
       {"text": page_title}
     ]
