+++ workbaskets/jinja2/workbaskets/edit-workbasket.jinja
@@ -1,0 +1,150 @@
+{% extends "layouts/layout.jinja" %}
+{% from "components/breadcrumbs/macro.njk" import govukBreadcrumbs %}
+
+{% set page_title %}Workbasket {{ request.session.workbasket.id }}{% if request.session.workbasket.title %} - {{ request.session.workbasket.title }}{% endif %}{% endset %}
+{% set page_subtitle %}What would you like to do?{% endset %}
+
+{% block breadcrumb %}
+  {% if request.GET.edit %}
+    {{ govukBreadcrumbs({
+      "items": [
+        {"text": "Home", "href": url("home")},
+        {"text": "Edit an existing workbasket", "href": url("workbaskets:select-workbasket")},
+        {"text": "Workbasket " ~ request.session.workbasket.id ~ " - Menu" }
+      ]
+    }) }}
+  {% else %}
+    {{ govukBreadcrumbs({
+      "items": [
+        {"text": "Home", "href": url("home")},
+        {"text": "Workbasket " ~ request.session.workbasket.id ~ " - Menu" }
+      ]
+    }) }}
+  {% endif %}
+{% endblock %}
+
+
+{% block content %}
+  <h1 class="govuk-heading-xl">
+    {{ page_title }}
+    <span class="govuk-caption-xl">
+      {{ page_subtitle }}
+    </span>
+  </h1>
+  <div class="govuk-grid-row">
+    <div class="govuk-grid-column-two-thirds">
+      <h2 class="govuk-heading-l">
+        Main menu
+      </h2>
+    </div>
+  </div>
+  <div class="govuk-grid-row">
+    <div class="govuk-grid-column-one-quarter">
+      <h2 class="govuk-heading-m">Additional codes</h2>
+      <ul class="govuk-list">
+        <li>
+          <a class="govuk-link" href="{{ url('additional_code-ui-create') }}">
+            Create a new additional code
+          </a>
+        </li>
+        <li>
+          <a class="govuk-link" href="{{ url("additional_code-ui-list") }}">
+            Find and edit additional codes
+          </a>
+        </li>
+      </ul>
+    </div>
+    <div class="govuk-grid-column-one-quarter">
+      <h2 class="govuk-heading-m">Certificates</h2>
+      <ul class="govuk-list">
+      <li>
+          <a class="govuk-link" href="{{ url('certificate-ui-create') }}">
+            Create a new certificate
+          </a>
+        </li>
+        <li>
+          <a class="govuk-link" href="{{ url("certificate-ui-list") }}">
+            Find and edit certificates
+          </a>
+        </li>
+      </ul>
+    </div>
+    <div class="govuk-grid-column-one-quarter">
+      <h2 class="govuk-heading-m">Footnotes</h2>
+      <ul class="govuk-list">
+        <li>
+          <a class="govuk-link" href="{{ url('footnote-ui-create') }}">
+            Create a new footnote
+          </a>
+        </li>
+        <li>
+          <a class="govuk-link" href="{{ url('footnote-ui-list') }}">
+            Find and edit footnotes
+          </a>
+        </li>
+      </ul>
+    </div>
+    <div class="govuk-grid-column-one-quarter">
+      <h2 class="govuk-heading-m">Geographical areas</h2>
+      <ul class="govuk-list">
+        <li>
+          <a class="govuk-link" href="{{ url('geo_area-ui-list') }}">
+            Find and edit geographical areas
+          </a>
+        </li>
+      </ul>
+    </div>
+  </div>
+  <div class="govuk-grid-row">
+    <div class="govuk-grid-column-one-quarter">
+      <h2 class="govuk-heading-m">Commodities</h2>
+      <ul class="govuk-list">
+        <li>
+          <a class="govuk-link" href="{{ url('commodities-import') }}">
+            Import commodities
+          </a>
+        </li>
+      </ul>
+    </div>
+    <div class="govuk-grid-column-one-quarter">
+      <h2 class="govuk-heading-m">Measures</h2>
+      <ul class="govuk-list">
+        <li>
+          <a class="govuk-link" href="{{ url('measure-ui-create', kwargs={"step": "start"}) }}">
+            Create new measure
+          </a>
+        </li>
+        <li>
+          <a class="govuk-link" href="{{ url('measure-ui-list') }}">
+            Find and edit measures
+          </a>
+        </li>
+      </ul>
+    </div>
+    <div class="govuk-grid-column-one-quarter">
+      <h2 class="govuk-heading-m">Quotas</h2>
+      <ul class="govuk-list">
+        <li>
+          <a class="govuk-link" href="{{ url('quota-ui-list') }}">
+            Find and edit quotas
+          </a>
+        </li>
+      </ul>
+    </div>
+    <div class="govuk-grid-column-one-quarter">
+      <h2 class="govuk-heading-m">Regulations</h2>
+      <ul class="govuk-list">
+        <li>
+          <a class="govuk-link" href="{{ url('regulation-ui-create') }}">
+            Create a new regulation
+          </a>
+        </li>
+        <li>
+          <a class="govuk-link" href="{{ url('regulation-ui-list') }}">
+            Find and edit regulations
+          </a>
+        </li>
+      </ul>
+    </div>
+  </div>
+{% endblock %}
