+++ workbaskets/jinja2/workbaskets/select-workbasket.jinja
@@ -8,7 +8,7 @@
 {% block breadcrumb %}
   {{ govukBreadcrumbs({
     "items": [
-      {"text": "Home", "href": url("index")},
+      {"text": "Home", "href": url("home")},
       {"text": page_title}
     ]
   }) }}
