+++ workbaskets/urls.py
@@ -22,9 +22,34 @@ ui_patterns = [
         name="workbasket-ui-create",
     ),
     path(
+        "preview-workbasket/",
+        ui_views.PreviewWorkbasketView.as_view(),
+        name="preview-workbasket",
+    ),
+    path(
+        "edit-workbasket/",
+        ui_views.EditWorkbasketView.as_view(),
+        name="edit-workbasket",
+    ),
+    path(
+        "review-workbasket/",
+        ui_views.ReviewWorkbasketView.as_view(),
+        name="review-workbasket",
+    ),
+    path(
+        "download",
+        ui_views.download_envelope,
+        name="workbasket-download",
+    ),
+    path(
         f"<pk>/",
         ui_views.WorkBasketDetail.as_view(),
         name="workbasket-ui-detail",
+    ),
+    path(
+        f"<pk>/confirm-create/",
+        ui_views.WorkBasketConfirmCreate.as_view(),
+        name="workbasket-ui-confirm-create",
     ),
     path(
         f"<pk>/submit/",
@@ -40,11 +65,6 @@ ui_patterns = [
         f"<pk>/delete-changes-done/",
         ui_views.WorkBasketDeleteChangesDone.as_view(),
         name="workbasket-ui-delete-changes-done",
-    ),
-    path(
-        "download",
-        ui_views.download_envelope,
-        name="workbasket-download",
     ),
 ]
